requirements.txt
opencv-python-headless==4.7.0.68
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
python-multipart
uvicorn
gunicorn
numpy>=1.18.5,<1.24.0
pandas>=1.1.4
Pillow>=7.1.2
requests>=2.23.0

tqdm>=4.41.0
PyYAML>=5.3.1
matplotlib>=3.2.2
seaborn>=0.11.0
scipy>=1.4.1
protobuf<4.21.3
requests>=2.23.0

fastapi
click
asgiref